fixed more issues and made more small improvements
